LayoutTests/media/video-poster-set-after-playback.html


<html>
<head>
    <title>Setting poster after playing</title>
    <script src=video-test.js></script>
    <script src=media-file.js></script>
    <script>

        window.addEventListener('load', async () => {
            failTestIn(15000);

            findMediaElement();
            video.src = findMediaFile("video", "content/test");

            video.play();
            await waitFor(video, 'playing');

            video.pause();
            await waitFor(video, 'pause');
            
            video.poster = 'content/green.png';
            
            await sleepFor(1000);
            
            if (window.internals)
                testExpected('internals.elementShouldDisplayPosterImage(video)', false);
            else
                consoleWrite("video should not be blank")

            endTest();
        });

    </script>
</head>
<body>
    <video controls muted></video>
</body>
</html>
